- Add local testnet test for successful doppelganger protection period - Update logging during doppelganger protection period

+++ scripts/tests/doppelganger_protection.sh
@@ -1,12 +1,19 @@
 #!/usr/bin/env bash
 
-cp ../local_testnet/vars.env ../local_testnet/vars.env.bkp
-cp ./vars.env ../local_testnet/vars.env
+BEHAVIOR=$1
+
+if [[ "$BEHAVIOR" != "success" ]] && [[ "$BEHAVIOR" != "failure" ]]; then
+    echo "Usage: doppelganger_protection.sh [success|failure]"
+    exit 1
+fi
+
+cp ./vars_$BEHAVIOR.env ./vars.env
+
 ../local_testnet/clean.sh
 
 echo "Starting ganache"
 
-../local_testnet/ganache_test_node.sh &
+../local_testnet/ganache_test_node.sh &> /dev/null &
 GANACHE_PID=$!
 
 # wait for ganache to start
@@ -21,7 +28,7 @@ cp -R $HOME/.lighthouse/local-testnet/node_1 $HOME/.lighthouse/local-testnet/nod
 
 echo "Starting bootnode"
 
-../local_testnet/bootnode.sh &
+../local_testnet/bootnode.sh &> /dev/null &
 BOOT_PID=$!
 
 # wait for the bootnode to start
@@ -29,44 +36,67 @@ sleep 5
 
 echo "Starting local beacon nodes"
 
-../local_testnet/beacon_node.sh $HOME/.lighthouse/local-testnet/node_1 9000 8000 &
+../local_testnet/beacon_node.sh $HOME/.lighthouse/local-testnet/node_1 9000 8000 &> /dev/null &
 BEACON_PID=$!
-../local_testnet/beacon_node.sh $HOME/.lighthouse/local-testnet/node_2 9100 8100 &
+../local_testnet/beacon_node.sh $HOME/.lighthouse/local-testnet/node_2 9100 8100 &> /dev/null &
 BEACON2_PID=$!
-../local_testnet/beacon_node.sh $HOME/.lighthouse/local-testnet/node_3 9200 8200 &
+../local_testnet/beacon_node.sh $HOME/.lighthouse/local-testnet/node_3 9200 8200 &> /dev/null &
 BEACON3_PID=$!
 
 echo "Starting local validator clients"
 
-../local_testnet/validator_client.sh $HOME/.lighthouse/local-testnet/node_1 http://localhost:8000 &
+../local_testnet/validator_client.sh $HOME/.lighthouse/local-testnet/node_1 http://localhost:8000 &> /dev/null &
 VALIDATOR_1_PID=$!
-../local_testnet/validator_client.sh $HOME/.lighthouse/local-testnet/node_2 http://localhost:8100 &
+../local_testnet/validator_client.sh $HOME/.lighthouse/local-testnet/node_2 http://localhost:8100 &> /dev/null &
 VALIDATOR_2_PID=$!
-../local_testnet/validator_client.sh $HOME/.lighthouse/local-testnet/node_3 http://localhost:8200 &
+../local_testnet/validator_client.sh $HOME/.lighthouse/local-testnet/node_3 http://localhost:8200 &> /dev/null &
 VALIDATOR_3_PID=$!
 
-echo "Waiting an epoch before starting the duplicate validator client"
-
-# Wait an epoch before starting a validator for doppelganger detection
+echo "Waiting an epoch before starting the next validator client"
 sleep 64
 
-echo "Starting the doppelganger validator client"
+if [ "$BEHAVIOR" == "failure" ]; then
 
-# Use same keys as keys from VC1, but connect to BN2
-# This process should not last longer than 2 epochs
-timeout 128 ../local_testnet/validator_client.sh $HOME/.lighthouse/local-testnet/node_1_doppelganger http://localhost:8100
-DOPPELGANGER_EXIT=$?
+    echo "Starting the doppelganger validator client"
 
-echo "Shutting down"
+    # Use same keys as keys from VC1, but connect to BN2
+    # This process should not last longer than 2 epochs
+    timeout 128 ../local_testnet/validator_client.sh $HOME/.lighthouse/local-testnet/node_1_doppelganger http://localhost:8100
+    DOPPELGANGER_EXIT=$?
 
-# Cleanup
-kill $BOOT_PID $BEACON_PID $BEACON2_PID $GANACHE_PID $VALIDATOR_1_PID $VALIDATOR_2_PID $VALIDATOR_3_PID $BEACON3_PID
-mv ../local_testnet/vars.env.bkp ../local_testnet/vars.env
+    echo "Shutting down"
 
-echo "Done"
+    # Cleanup
+    kill $BOOT_PID $BEACON_PID $BEACON2_PID $GANACHE_PID $VALIDATOR_1_PID $VALIDATOR_2_PID $VALIDATOR_3_PID $BEACON3_PID
+    rm ./vars.env
 
-if [ $DOPPELGANGER_EXIT -eq 124 ]; then
-    exit 1
+    echo "Done"
+
+    if [ $DOPPELGANGER_EXIT -eq 124 ]; then
+        exit 1
+    fi
+fi
+
+if [ "$BEHAVIOR" == "success" ]; then
+
+    echo "Starting the last validator client"
+
+    # This process should last longer than 3 epochs
+    timeout 192 ../local_testnet/validator_client.sh $HOME/.lighthouse/local-testnet/node_4 http://localhost:8100
+
+    DOPPELGANGER_EXIT=$?
+
+    echo "Shutting down"
+
+    # Cleanup
+    kill $BOOT_PID $BEACON_PID $BEACON2_PID $GANACHE_PID $VALIDATOR_1_PID $VALIDATOR_2_PID $VALIDATOR_3_PID $BEACON3_PID
+    rm ./vars.env
+
+    echo "Done"
+
+    if [ $DOPPELGANGER_EXIT -ne 124 ]; then
+        exit 1
+    fi
 fi
 
 exit 0
